Move boundary details query inside <Map>
When the query was outside of the <Map> component, it would cause
re-render of the map each time the query data was updated.
